Completed Part 4
Went back and remembered that I should be putting the configurations at the bottom, underneath the main. Then I remembered that if I wanted to to make the device to change modes after a certain amount of seconds, a counter would need to be initialized and ran throughout the code.